refactor : 두 API에서 공통으로 사용되는 getOperationTimeOfWeek 함수를 파일로 분리
